requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
tavily-python
